Wrongly moved boltzmann.c to boltzmann_t1.c, the reason is that I am not confident enough to retrieve an old file from git yet.

+++ c/boltzmann.h
@@ -30,7 +30,7 @@ struct node {
 };
 
 struct layer {
-    int num_nodes;
+    size_t num_nodes;
     int kind;
     struct node *node;
 };
